Updated RES model. 3 body final state phase space added, code copied from SHERPA, can use delta res for 2 hadron final state or flat distribution. Currently no piminus are in the final state, also inclusive cross section is off by a factor of 2.

+++ src/Achilles/fortran/nuclear_model.f90
@@ -53,8 +53,8 @@ module nuclear_model
             integer(c_size_t), intent(in), value :: nin, nout, len_ff, nspin, nlorentz
             complex(c_double_complex), dimension(len_ff), intent(in) :: ff
             type(fourvector) :: qvec
-            integer(c_int), dimension(nin), intent(in) :: pids_in
-            integer(c_int), dimension(nout), intent(in) :: pids_out
+            integer(c_long), dimension(nin), intent(in) :: pids_in
+            integer(c_long), dimension(nout), intent(in) :: pids_out
             type(fourvector), dimension(nin), intent(in) :: mom_in
             type(fourvector), dimension(nout), intent(in) :: mom_out
             complex(c_double_complex), dimension(nlorentz, nspin), intent(out) :: cur
